1. Converted Elliposoid code to double. 2. Matrix.h bug fix. Floats are converted to template 3. Elliptic sampling bug fix 4. BDPT bug fix. Current code does direct sampling even if the flag is set to false. Fixed it 5. Ellipsoid structure is extended to contain shading normal. EarlyTriangleReject is added based on that.

+++ include/mitsuba/core/fwd.h
@@ -159,7 +159,7 @@ typedef TAABB<Point4>         AABB4;
 typedef TRay<Point, Vector>   Ray;
 typedef TRay<Point2, Vector2> Ray2;
 template <typename PointType, typename LengthType> struct TEllipsoid;
-typedef TEllipsoid<Point3f, float> Ellipsoid;
+typedef TEllipsoid<Point3d, double> Ellipsoid;
 struct RayPacket4;
 struct RayInterval4;
 struct Intersection4;
